10. REPL Mode Reference

10. REPL Mode Reference

This section describes available commands in the Pkg REPL. The REPL mode is mostly meant for interactive use, and for non-interactive use it is recommended to use the "API mode", see API Reference.

package commands

add — REPL command .

build — REPL command .

develop — REPL command .

free — REPL command .

generate — REPL command .

pin — REPL command .

remove — REPL command .

test — REPL command .

update — REPL command .

registry commands

registry add — REPL command .

registry remove — REPL command .

registry status — REPL command .

registry update — REPL command .

Other commands

activate — REPL command .

gc — REPL command .

help — REPL command .

instantiate — REPL command .

precompile — REPL command .

resolve — REPL command .

status — REPL command .